Add homepage and styling files
adds the first version of the home page (HomePage.html) and the style sheet (styles.css) that goes with it. The title of the employee management system and a button to show all workers are on the home page. The look of the page's parts is set by the CSS file. In the future, more features will be added to make the programme work better and look better.